requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
importlib-metadata==2.1.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
streamlit
matplotlib
